Move classproperty to a dedicated private module
This was declared in `globus_sdk.utils`. Moving it to
`globus_sdk._classproperty` has several benefits:
- it is easier to read because the "boundary" around it is at the
  file-level (previously, it was a chunk of a file)
- it makes `utils` smaller and therefore also easier to read
- it makes it even more clear that this is an internal component --
  historically, some users have pulled in `globus_sdk.utils` and used
  its components
